Stick to all Noto Sans fonts for i18n example
A hand curated list of mixed styles *could* be a good thing, but it
wasn't carefully curated it was just kind of cobbled together with
anything that worked. The result was too many downloads and a chaotic
feel. I like examples that demonstrate one thing at a time, and I think
sticking to one broad family of fonts will help this example.